+++ CHANGELOG.md
@@ -52,6 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `compas_rhino.draw_circles` and the equivalent Artist method
 - Add class functions to `compas.datastructures.VolMesh`.
 - Added `face_neighborhood` class function to `compas.datastructures.Mesh`.
+- Added `get_face_attributes_all` to `compas.datastructures._mixins.attributes`.
+- Added `get_faces_attributes_all` to `compas.datastructures._mixins.attributes`.
 - Added `compas.remote` package for making HTTP based Remote Procedure Calls.
 
 ### Changed